A+C: Add information about my contributions
As requested by @cznic add information about me and my company to
AUTHORS and CONTRIBUTORS.

Btw, I'm not a lawyer and maybe there are country differences, but I
used to think that:

- author is the person who actually creates the work.
- copyright holder is who owns the copyright on that work (i.e. the
  person, or company he is working for or someone else).

The `AUTHORS` file however gets it the other way - it says to put there
information about who is "copyright holder" which goes counterwise above
classification.

----

I understand however that the exact A+C approach is taken from the Go
project so let's also continue following it here.